feat: adapt index to use AppService
